+++ tests/testthat/test_covariance_plot.R
@@ -1,3 +1,5 @@
+# h_get_timepoint_vars ----
+
 test_that("h_get_timepoint_vars works in simple example", {
   vcov_matrix <- matrix(
     c(49, 24, 12, 23, 24, 35, 11, 20, 12, 11, 24, 14, 23, 20, 14, 107),
@@ -7,7 +9,7 @@ test_that("h_get_timepoint_vars works in simple example", {
       c("VIS1", "VIS2", "VIS3", "VIS4")
     )
   )
-  result <- expect_silent(h_get_timepoint_vars(vcov_matrix, string = "VIS"))
+  result <- expect_silent(h_get_timepoint_vars(vcov_matrix, time_prefix = "VIS"))
   expected <- list(
     row_time = c(1, 1, 2, 1, 2, 3, 1, 2, 3, 4),
     col_time = c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
@@ -15,10 +17,10 @@ test_that("h_get_timepoint_vars works in simple example", {
   expect_identical(result, expected)
 })
 
-test_that("h_get_timepoint_vars works as expected with string specified", {
+test_that("h_get_timepoint_vars works as expected with time_prefix specified", {
   data_wide <- maditr::dcast(mmrm_test_data, USUBJID + ARMCD ~ AVISIT, value.var = "FEV1")
   data_cov <- cov(data_wide[, 3:ncol(data_wide)], use = "pairwise.complete.obs", method = "pearson")
-  result <- h_get_timepoint_vars(vcov_matrix = data_cov, string = "VIS")
+  result <- h_get_timepoint_vars(vcov_matrix = data_cov, time_prefix = "VIS")
   expected <- list(
     row_time = c(1, 1, 2, 1, 2, 3, 1, 2, 3, 4),
     col_time = c(1, 2, 2, 3, 3, 3, 4, 4, 4, 4)
@@ -26,13 +28,13 @@ test_that("h_get_timepoint_vars works as expected with string specified", {
   expect_identical(result, expected)
 })
 
-test_that("h_get_timepoint_vars returns error when string not specified", {
+test_that("h_get_timepoint_vars returns error when time_prefix not specified", {
   data_wide <- maditr::dcast(mmrm_test_data, USUBJID + ARMCD ~ AVISIT, value.var = "FEV1")
   data_cov <- cov(data_wide[, 3:ncol(data_wide)], use = "pairwise.complete.obs", method = "pearson")
   expect_error(h_get_timepoint_vars(vcov_matrix = data_cov))
 })
 
-test_that("h_get_timepoint_vars works as expected with string argument not needed", {
+test_that("h_get_timepoint_vars works as expected with time_prefix argument not needed", {
   data_wide <- maditr::dcast(mmrm_test_data, USUBJID + ARMCD ~ AVISIT, value.var = "FEV1")
   data_cov <- cov(data_wide[, 3:ncol(data_wide)], use = "pairwise.complete.obs", method = "pearson")
   row.names(data_cov) <- c(1, 2, 3, 4)
@@ -45,7 +47,9 @@ test_that("h_get_timepoint_vars works as expected with string argument not neede
   expect_identical(result, expected)
 })
 
-test_that("h_vectorization works as expected with string specified", {
+# h_vectorization ----
+
+test_that("h_vectorization works as expected with time_prefix specified", {
   vcov_matrix <- matrix(
     c(49, 12, 12, 23),
     nrow = 2, ncol = 2,
@@ -54,10 +58,10 @@ test_that("h_vectorization works as expected with string specified", {
       c("VIS1", "VIS2")
     )
   )
-  result <- h_vectorization(vcov_matrix, string = "VIS")
+  result <- h_vectorization(vcov_matrix, time_prefix = "VIS")
   expected <- data.frame(
     Vect = c(49, 12, 23),
-    time_point_distribution = c(0, 1, 0),
+    time_diff = c(0, 1, 0),
     lag = c(0, 1, 0),
     rank_row = c(1, 1, 2),
     rank_col = c(1, 2, 2)
@@ -65,7 +69,7 @@ test_that("h_vectorization works as expected with string specified", {
   expect_identical(result, expected)
 })
 
-test_that("h_vectorization works as expected when string not specified", {
+test_that("h_vectorization works as expected when time_prefix not specified", {
   vcov_matrix <- matrix(
     c(49, 12, 12, 23),
     nrow = 2, ncol = 2,
@@ -77,7 +81,7 @@ test_that("h_vectorization works as expected when string not specified", {
   result <- h_vectorization(vcov_matrix)
   expected <- data.frame(
     Vect = c(49, 12, 23),
-    time_point_distribution = c(0, 1, 0),
+    time_diff = c(0, 1, 0),
     lag = c(0, 1, 0),
     rank_row = c(1, 1, 2),
     rank_col = c(1, 2, 2)
@@ -85,7 +89,9 @@ test_that("h_vectorization works as expected when string not specified", {
   expect_identical(result, expected)
 })
 
-test_that("g_covarariance works as expected as expected with no xlab specified", {
+# g_covariance ----
+
+test_that("g_covariance works as expected as expected with defaults", {
   vcov_matrix <- matrix(
     c(49, 12, 12, 23),
     nrow = 2, ncol = 2,
@@ -95,10 +101,10 @@ test_that("g_covarariance works as expected as expected with no xlab specified",
     )
   )
   result <- expect_silent(g_covariance(vcov_matrix))
-  vdiffr::expect_doppelganger("g_covariance plot no xlab specified", result)
+  vdiffr::expect_doppelganger("g_covariance plot with defaults", result)
 })
 
-test_that("g_covarariance works as expected as expected with string specified", {
+test_that("g_covariance works as expected as expected with time_prefix specified", {
   vcov_matrix <- matrix(
     c(49, 12, 12, 23),
     nrow = 2, ncol = 2,
@@ -107,6 +113,19 @@ test_that("g_covarariance works as expected as expected with string specified", 
       c("VIS1", "VIS2")
     )
   )
-  result <- expect_silent(g_covariance(vcov_matrix, string = "VIS"))
-  vdiffr::expect_doppelganger("g_covariance plot with string specified", result)
+  result <- expect_silent(g_covariance(vcov_matrix, time_prefix = "VIS"))
+  vdiffr::expect_doppelganger("g_covariance plot with time_prefix specified", result)
+})
+
+test_that("g_covariance works as expected as expected with time difference", {
+  vcov_matrix <- matrix(
+    c(49, 12, 12, 23),
+    nrow = 2, ncol = 2,
+    dimnames = list(
+      c(1, 2),
+      c(1, 2)
+    )
+  )
+  result <- expect_silent(g_covariance(vcov_matrix, x_var = "time_diff"))
+  vdiffr::expect_doppelganger("g_covariance plot with time diff", result)
 })